+++ myapp/lib/부대코드입력.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:get/get.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(UnitCode());
 const primaryColor = Color(0xFFF7CBD4);
 
-class MyApp extends StatelessWidget {
+class UnitCode extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     //final appTitle = 'sign up page';
@@ -17,11 +18,13 @@ class MyApp extends StatelessWidget {
       home: Scaffold(
         appBar: AppBar(
           backgroundColor: primaryColor,
-          title: Text('사용자 개인정보 입력페이지'),
+          title: Text('부대 코드 입력 페이지'),
           automaticallyImplyLeading: true,
           leading: IconButton(
             icon: Icon(Icons.arrow_back_ios_new),
-            onPressed: () {},
+            onPressed: () {
+              Get.back();
+            },
           ),
         ),
         body: MyCustomForm(),
@@ -48,7 +51,7 @@ class MyCustomFormState extends State<MyCustomForm> {
     // Build a Form widget using the _formKey created above.
     return Column(
       children: [
-        SizedBox(height: 120),
+        const SizedBox(height: 170),
         Form(
           key: _formKey,
           child: Container(
@@ -65,83 +68,18 @@ class MyCustomFormState extends State<MyCustomForm> {
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.center,
               children: <Widget>[
-                SizedBox(height: 40),
+                const SizedBox(height: 100),
                 TextFormField(
                   decoration: const InputDecoration(
                     border: OutlineInputBorder(),
                     labelStyle: TextStyle(fontWeight: FontWeight.bold),
                     icon: const Icon(Icons.person),
-                    hintText: '- 를 포함하여 입력해 주세요.',
-                    labelText: '아이디(군번)',
+                    hintText: '부대 코드를 입력하세요.',
+                    labelText: '코드 입력',
                   ),
                   validator: (value) {
                     if (value!.isEmpty) {
-                      return '군번을 입력해주세요';
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 10),
-
-                TextFormField(
-                  decoration: const InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelStyle: TextStyle(fontWeight: FontWeight.bold),
-                    icon: const Icon(Icons.password),
-                    hintText: '비밀번호를 입력해주세요',
-                    labelText: '비밀번호',
-                  ),
-                  validator: (value) {
-                    if (value!.isEmpty) {
-                      return '비밀번호를 입력해주세요';
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 10),
-                TextFormField(
-                  decoration: const InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelStyle: TextStyle(fontWeight: FontWeight.bold),
-                    icon: const Icon(Icons.near_me),
-                    hintText: '이름을 입력해 주세요  ex) 홍길동',
-                    labelText: '이름',
-                  ),
-                  validator: (value) {
-                    if (value!.isEmpty) {
-                      return '이름을 입력해주세요';
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 10),
-                TextFormField(
-                  decoration: const InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelStyle: TextStyle(fontWeight: FontWeight.bold),
-                    icon: const Icon(Icons.person_add_alt_rounded),
-                    hintText: '계급을 입력해 주세요  ex) 일병',
-                    labelText: '계급',
-                  ),
-                  validator: (value) {
-                    if (value!.isEmpty) {
-                      return '계급을 입력해주세요';
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 10),
-                TextFormField(
-                  decoration: const InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelStyle: TextStyle(fontWeight: FontWeight.bold),
-                    icon: const Icon(Icons.email),
-                    hintText: '이메일 주소를 입력해 주세요  ex)osam2021@naver.com',
-                    labelText: '이메일 주소',
-                  ),
-                  validator: (value) {
-                    if (value!.isEmpty) {
-                      return '이메일주소를 입력해주세요';
+                      return '부대 코드를 입력하세요.';
                     }
                     return null;
                   },
